refactor del post action adn used hook
